requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
langchain-community
duckduckgo-search
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
